housing_livelihood_agent/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1

datasets==2.14.7
pyarrow==14.0.1
huggingface-hub==0.22.2

tenacity>=8.2.2
prometheus-fastapi-instrumentator>=6.1.0

scikit-learn>=1.5.0
pandas>=2.2.2
folium>=0.17.0

langchain==0.1.16
langchain-community==0.0.34
transformers==4.39.3  # compatible with HF Hub 0.22.2
torch>=2.2.0
accelerate>=0.30.0
